+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.2.0-pl] - 2026-09-22
+
+### Added
+- Button description under the title
+- Per-button background color
+- General settings for the default background and text color. With no background, buttons are gray
+
 ## [1.1.1-pl] - 2026-09-22
 
 ### Fixed